Move endpoint interface to endpoint package. Also add comment.
